feat: extract shared components.
